% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{critical_pressure}
\alias{critical_pressure}
\title{Critical Pressure}
\usage{
critical_pressure(Temperature = 25, type = "saturated")
}
\arguments{
\item{Temperature}{the temperature of seawater. default: 25 degree C}

\item{type}{should be either "saturated" or "full". default: "saturated"}
}
\value{
seawater pressure at critical curve
}
\description{
calculate the critical pressure (bar) of seawater at given temperature using the equation proposed by Driesner and Heinrich (2007)
}
\examples{
critical_pressure(Temperature = 100, type="saturated")
The output is: 1.901956 with an unit of bar
}
