% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{d18O_temperature}
\alias{d18O_temperature}
\title{carbonate veins}
\usage{
d18O_temperature(d18O = 20, mineral, ref = 0)
}
\arguments{
\item{d18O}{the d18O value of the carbonate veins. default: 20}

\item{mineral}{the mineral type of the carbonate veins.}

\item{ref}{the d18O of precipitation fluid. default: 0}
}
\value{
Precipitation temperature of carbonate veins
}
\description{
calculate the precipitation temperature of carbonate veins
}
\examples{
d18O_temperature(d18O = 20, mineral = "dolomite", ref = -1)
The output is: 88.6529 with an unit of degree C
}
