% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dissolution_constant}
\alias{dissolution_constant}
\title{data frame calculation}
\usage{
dissolution_constant(Temperature = 25, mineral)
}
\arguments{
\item{Temperature}{temperature condition. default: 25 degree C}

\item{mineral}{the specified minerals: anhydrite, calcite, barite, celestite}
}
\value{
dissolution constant of the specified mineral
}
\description{
calculate the dissolution constant of a specified mineral at given temperature
}
\examples{
dissolution_constant(Temperature = 100, mineral = "anhydrite")
The output is: -2.497612
}
