% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{Henry_constant}
\alias{Henry_constant}
\title{Henry's constants}
\usage{
Henry_constant(Temperature = 25, gas = "N2")
}
\arguments{
\item{Temperature}{temperature condition. default: 25 degree C}

\item{gas}{the gas species. default: "N2"}
}
\value{
KH constant of a gas at given temperature
}
\description{
calculate the Henry's constants
}
\examples{
Henry_constant(Temperature = 100, gas = "Ar")
The output is: 64421.35
}
