% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{Horita}
\alias{Horita}
\title{data frame calculation}
\usage{
Horita(dC)
}
\arguments{
\item{dC}{ΔCO2-CH4}
}
\value{
fractionation of δ13C between CO2 and CH4
}
\description{
calculate the equilibration temperature based on the stable carbon fractionation between CO2 and CH4
}
\examples{
Horita (dC = 50)
The output is: 212 with  unit of degree C
}
