% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{Horita_T}
\alias{Horita_T}
\title{data frame calculation}
\usage{
Horita_T(Temperature = 25)
}
\arguments{
\item{Temperature}{temperature condition. default: 25 degree C}
}
\value{
fractionation of δ13C between CO2 and CH4
}
\description{
calculate stable carbon fractionation between CO2 and CH4
}
\examples{
Horita_T (Temperature = 50)
The output is: 61.45526
}
