% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{quartz_sol}
\alias{quartz_sol}
\title{Quartz solubility}
\usage{
quartz_sol(Temperature = 25, S = 35, ref = "vonDamm1991")
}
\arguments{
\item{Temperature}{the temperature of seawater. default: 25 degree C}

\item{S}{the salinity of seawater. default: 35}

\item{ref}{the reference that used to calculate the quartz solubility. "vonDamm1991", "Gunnarsson2000". default: "vonDamm1991"}
}
\value{
dissolved SiO2 concentration in the fluid that equilibrate with quartz at giving condition
}
\description{
calculate the quartz solubility
}
\examples{
quartz_sol(Temperature = 100, S = 35, ref = "vonDamm1991")
The output is: 1.003644 with an unit of mM
}
