% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{REE_normalization}
\alias{REE_normalization}
\title{data frame calculation}
\usage{
REE_normalization(data, method = "CI")
}
\arguments{
\item{data}{the raw data that need to be normalized}

\item{method}{the normalization reference. default: "CI"}
}
\value{
a data frame with REEs data normalized by CI chondrite or PAAS
}
\description{
Normalize the REEs data
}
\examples{
REE_normalization (data = temp, method = "PAAS")
}
