% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{seawater_density}
\alias{seawater_density}
\title{value calculation}
\usage{
seawater_density(Temperature = 25, P = 1, S = 35)
}
\arguments{
\item{Temperature}{the temperature of seawater. default: 25 degree C}

\item{P}{the pressure condition of seawater. default: 1 bar}

\item{S}{seawater salinity. default: 35}
}
\value{
seawater density at this P-T-S condition
}
\description{
calculate seawater density as a function of T, P, and salinity using the equations proposed by Safarov et al., (2012,2013)
}
\examples{
seawater_density(Temperature = 100, P = 20, S = 35)
The output is: 0.9846648 with an unit of g cm-3
}
